Fix packager when runtime.js was autogenerated and already exists in output directory.

+++ sdks/wasm/packager.cs
@@ -292,12 +292,16 @@ class Driver {
 			File.Delete (runtime_js);
 			File.Copy (runtimeTemplate, runtime_js);
 		} else {
-			if (File.Exists(runtime_js)) {
-				CopyFile (runtimeTemplate, runtime_js, CopyType.IfNewer, $"runtime template <{runtimeTemplate}> ");
-			} else {
-				var runtime_gen = "\nvar Module = {\n\tonRuntimeInitialized: function () {\n\t\tMONO.mono_load_runtime_and_bcl (\n\t\tconfig.vfs_prefix,\n\t\tconfig.deploy_prefix,\n\t\tconfig.enable_debugging,\n\t\tconfig.file_list,\n\t\tfunction () {\n\t\t\tconfig.add_bindings ();\n\t\t\tApp.init ();\n\t\t}\n\t)\n\t},\n};";
-				File.Delete (runtime_js);
-				File.WriteAllText (runtime_js, runtime_gen);
+            if (File.Exists(runtime_js) && (File.Exists(runtimeTemplate))) {
+                CopyFile (runtimeTemplate, runtime_js, CopyType.IfNewer, $"runtime template <{runtimeTemplate}> ");
+            } else {
+                if (File.Exists(runtimeTemplate))
+                    CopyFile(runtimeTemplate, runtime_js, CopyType.Always, $"runtime template <{runtimeTemplate}> ");
+                else {
+                    var runtime_gen = "\nvar Module = {\n\tonRuntimeInitialized: function () {\n\t\tMONO.mono_load_runtime_and_bcl (\n\t\tconfig.vfs_prefix,\n\t\tconfig.deploy_prefix,\n\t\tconfig.enable_debugging,\n\t\tconfig.file_list,\n\t\tfunction () {\n\t\t\tconfig.add_bindings ();\n\t\t\tApp.init ();\n\t\t}\n\t)\n\t},\n};";
+                    File.Delete(runtime_js);
+                    File.WriteAllText(runtime_js, runtime_gen);
+                }
 			}
 		}
